+++ enterprise/cmd/frontend/auth/httpheader/testproxy.go
@@ -15,10 +15,11 @@ import (
 )
 
 var (
-	addr       = flag.String("addr", ":4080", "HTTP listen address")
-	urlStr     = flag.String("url", "http://localhost:3080", "proxy origin URL (Sourcegraph HTTP/HTTPS URL)") // CI:LOCALHOST_OK
-	username   = flag.String("username", os.Getenv("USER"), "username to report to Sourcegraph")
-	httpHeader = flag.String("header", "X-Forwarded-User", "name of HTTP header to add to request")
+	addr           = flag.String("addr", ":4080", "HTTP listen address")
+	urlStr         = flag.String("url", "http://localhost:3080", "proxy origin URL (Sourcegraph HTTP/HTTPS URL)") // CI:LOCALHOST_OK
+	username       = flag.String("username", os.Getenv("USER"), "username to report to Sourcegraph")
+	usernamePrefix = flag.String("usernamePrefix", "", "prefix to place in front of username in the auth header value")
+	httpHeader     = flag.String("header", "X-Forwarded-User", "name of HTTP header to add to request")
 )
 
 func main() {
@@ -35,11 +36,12 @@ func main() {
 	if *httpHeader == "" {
 		log.Fatal("Error: No -header specified.")
 	}
-	log.Printf(`Listening on %s, forwarding requests to %s with added header "%s: %s"`, *addr, url, *httpHeader, *username)
+	headerVal := *usernamePrefix + *username
+	log.Printf(`Listening on %s, forwarding requests to %s with added header "%s: %s"`, *addr, url, *httpHeader, headerVal)
 	p := httputil.NewSingleHostReverseProxy(url)
 	log.Fatalf("Server error: %s.", http.ListenAndServe(*addr, &httputil.ReverseProxy{
 		Director: func(r *http.Request) {
-			r.Header.Set(*httpHeader, *username)
+			r.Header.Set(*httpHeader, headerVal)
 			r.Host = url.Host
 			p.Director(r)
 		},