+++ schema/site_stringdata.go
@@ -1096,6 +1096,12 @@ const SiteSchemaJSON = `{
           "description":
             "The name (case-insensitive) of an HTTP header whose value is taken to be the username of the client requesting the page. Set this value when using an HTTP proxy that authenticates requests, and you don't want the extra configurability of the other authentication methods.",
           "type": "string"
+        },
+        "stripUsernameHeaderPrefix": {
+          "description":
+            "The prefix that precedes the username portion of the HTTP header specified in ` + "`" + `usernameHeader` + "`" + `. If specified, the prefix will be stripped from the header value and the remainder will be used as the username. For example, if using Google Identity-Aware Proxy (IAP) with Google Sign-In, set this value to ` + "`" + `accounts.google.com:` + "`" + `.",
+          "type": "string",
+          "examples": ["accounts.google.com:"]
         }
       }
     },